Refactor StreamResponse, PublisherStream. - Embedded the distinction between `message` and `result` in test cases.
