+++ firebase-functions/src/main/java/com/google/firebase/functions/FirebaseFunctions.kt
@@ -30,14 +30,10 @@ import com.google.firebase.functions.FirebaseFunctionsException.Code.Companion.f
 import com.google.firebase.functions.FirebaseFunctionsException.Companion.fromResponse
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
-import java.io.BufferedReader
 import java.io.IOException
-import java.io.InputStream
-import java.io.InputStreamReader
 import java.io.InterruptedIOException
 import java.net.MalformedURLException
 import java.net.URL
-import java.util.concurrent.ConcurrentLinkedQueue
 import java.util.concurrent.Executor
 import javax.inject.Named
 import okhttp3.Call
@@ -50,8 +46,6 @@ import okhttp3.Response
 import org.json.JSONException
 import org.json.JSONObject
 import org.reactivestreams.Publisher
-import org.reactivestreams.Subscriber
-import org.reactivestreams.Subscription
 
 /** FirebaseFunctions lets you call Cloud Functions for Firebase. */
 public class FirebaseFunctions
@@ -323,253 +317,24 @@ internal constructor(
     data: Any?,
     options: HttpsCallOptions
   ): Publisher<StreamResponse> {
+    val url = getURL(name)
+    Preconditions.checkNotNull(url, "url cannot be null")
     val task =
       providerInstalled.task.continueWithTask(executor) {
         contextProvider.getContext(options.limitedUseAppCheckTokens)
       }
 
-    return PublisherStream(getURL(name), data, options, client, serializer, task, executor)
+    return PublisherStream(url, data, options, client, serializer, task, executor)
   }
 
   internal fun stream(url: URL, data: Any?, options: HttpsCallOptions): Publisher<StreamResponse> {
+    Preconditions.checkNotNull(url, "url cannot be null")
     val task =
       providerInstalled.task.continueWithTask(executor) {
         contextProvider.getContext(options.limitedUseAppCheckTokens)
       }
 
     return PublisherStream(url, data, options, client, this.serializer, task, executor)
-  }
-
-  private class PublisherStream(
-    private val url: URL,
-    private val data: Any?,
-    private val options: HttpsCallOptions,
-    private val client: OkHttpClient,
-    private val serializer: Serializer,
-    private val contextTask: Task<HttpsCallableContext?>,
-    private val executor: Executor
-  ) : Publisher<StreamResponse> {
-
-    private val subscribers = ConcurrentLinkedQueue<Subscriber<in StreamResponse>>()
-    private var activeCall: Call? = null
-
-    override fun subscribe(subscriber: Subscriber<in StreamResponse>) {
-      subscribers.add(subscriber)
-      subscriber.onSubscribe(
-        object : Subscription {
-          override fun request(n: Long) {
-            startStreaming()
-          }
-
-          override fun cancel() {
-            cancelStream()
-            subscribers.remove(subscriber)
-          }
-        }
-      )
-    }
-
-    private fun startStreaming() {
-      contextTask.addOnCompleteListener(executor) { contextTask ->
-        if (!contextTask.isSuccessful) {
-          notifyError(
-            FirebaseFunctionsException(
-              "Error retrieving context",
-              FirebaseFunctionsException.Code.INTERNAL,
-              null,
-              contextTask.exception
-            )
-          )
-          return@addOnCompleteListener
-        }
-
-        Preconditions.checkNotNull(url, "url cannot be null")
-        val context = contextTask.result
-        val callClient = options.apply(client)
-        val requestBody =
-          RequestBody.create(
-            MediaType.parse("application/json"),
-            JSONObject(mapOf("data" to serializer.encode(data))).toString()
-          )
-        val requestBuilder =
-          Request.Builder().url(url).post(requestBody).header("Accept", "text/event-stream")
-        applyCommonConfiguration(requestBuilder, context)
-        val request = requestBuilder.build()
-        val call = callClient.newCall(request)
-        activeCall = call
-
-        call.enqueue(
-          object : Callback {
-            override fun onFailure(call: Call, e: IOException) {
-              val message: String
-              val code: FirebaseFunctionsException.Code
-              if (e is InterruptedIOException) {
-                message = FirebaseFunctionsException.Code.DEADLINE_EXCEEDED.name
-                code = FirebaseFunctionsException.Code.DEADLINE_EXCEEDED
-              } else {
-                message = FirebaseFunctionsException.Code.INTERNAL.name
-                code = FirebaseFunctionsException.Code.INTERNAL
-              }
-              notifyError(FirebaseFunctionsException(message, code, null, e))
-            }
-
-            override fun onResponse(call: Call, response: Response) {
-              validateResponse(response)
-              val bodyStream = response.body()?.byteStream()
-              if (bodyStream != null) {
-                processSSEStream(bodyStream)
-              } else {
-                notifyError(
-                  FirebaseFunctionsException(
-                    "Response body is null",
-                    FirebaseFunctionsException.Code.INTERNAL,
-                    null
-                  )
-                )
-              }
-            }
-          }
-        )
-      }
-    }
-
-    private fun cancelStream() {
-      activeCall?.cancel()
-      notifyError(
-        FirebaseFunctionsException(
-          "Stream was canceled",
-          FirebaseFunctionsException.Code.CANCELLED,
-          null
-        )
-      )
-    }
-
-    private fun applyCommonConfiguration(
-      requestBuilder: Request.Builder,
-      context: HttpsCallableContext?
-    ) {
-      context?.authToken?.let { requestBuilder.header("Authorization", "Bearer $it") }
-      context?.instanceIdToken?.let { requestBuilder.header("Firebase-Instance-ID-Token", it) }
-      context?.appCheckToken?.let { requestBuilder.header("X-Firebase-AppCheck", it) }
-    }
-
-    private fun processSSEStream(inputStream: InputStream) {
-      BufferedReader(InputStreamReader(inputStream)).use { reader ->
-        try {
-          reader.lineSequence().forEach { line ->
-            val dataChunk =
-              when {
-                line.startsWith("data:") -> line.removePrefix("data:")
-                line.startsWith("result:") -> line.removePrefix("result:")
-                else -> return@forEach
-              }
-            try {
-              val json = JSONObject(dataChunk)
-              when {
-                json.has("message") ->
-                  serializer.decode(json.opt("message"))?.let {
-                    notifyData(StreamResponse.Message(data = it))
-                  }
-                json.has("error") -> {
-                  serializer.decode(json.opt("error"))?.let {
-                    notifyError(
-                      FirebaseFunctionsException(
-                        it.toString(),
-                        FirebaseFunctionsException.Code.INTERNAL,
-                        it
-                      )
-                    )
-                  }
-                }
-                json.has("result") -> {
-                  serializer.decode(json.opt("result"))?.let {
-                    notifyData(StreamResponse.Result(data = it))
-                    notifyComplete()
-                  }
-                  return
-                }
-              }
-            } catch (e: Throwable) {
-              notifyError(
-                FirebaseFunctionsException(
-                  "Invalid JSON: $dataChunk",
-                  FirebaseFunctionsException.Code.INTERNAL,
-                  e
-                )
-              )
-            }
-          }
-          notifyError(
-            FirebaseFunctionsException(
-              "Stream ended unexpectedly without completion",
-              FirebaseFunctionsException.Code.INTERNAL,
-              null
-            )
-          )
-        } catch (e: Exception) {
-          notifyError(
-            FirebaseFunctionsException(
-              e.message ?: "Error reading stream",
-              FirebaseFunctionsException.Code.INTERNAL,
-              e
-            )
-          )
-        }
-      }
-    }
-
-    private fun notifyData(data: StreamResponse?) {
-      for (subscriber in subscribers) {
-        subscriber.onNext(data)
-      }
-    }
-
-    private fun notifyError(e: FirebaseFunctionsException) {
-      for (subscriber in subscribers) {
-        subscriber.onError(e)
-      }
-      subscribers.clear()
-    }
-
-    private fun notifyComplete() {
-      for (subscriber in subscribers) {
-        subscriber.onComplete()
-      }
-      subscribers.clear()
-    }
-
-    private fun validateResponse(response: Response) {
-      if (response.isSuccessful) return
-
-      val htmlContentType = "text/html; charset=utf-8"
-      val trimMargin: String
-      if (response.code() == 404 && response.header("Content-Type") == htmlContentType) {
-        trimMargin = """URL not found. Raw response: ${response.body()?.string()}""".trimMargin()
-        throw FirebaseFunctionsException(
-          trimMargin,
-          FirebaseFunctionsException.Code.fromHttpStatus(response.code()),
-          null
-        )
-      }
-
-      val text = response.body()?.string() ?: ""
-      val error: Any?
-      try {
-        val json = JSONObject(text)
-        error = serializer.decode(json.opt("error"))
-      } catch (e: Throwable) {
-        throw FirebaseFunctionsException(
-          "${e.message} Unexpected Response:\n$text ",
-          FirebaseFunctionsException.Code.INTERNAL,
-          e
-        )
-      }
-      throw FirebaseFunctionsException(
-        error.toString(),
-        FirebaseFunctionsException.Code.INTERNAL,
-        error
-      )
-    }
   }
 
   public companion object {
